# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Software Department Portfolio Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cycles:
    get:
      operationId: listCycles
      tags: [cycles]
      summary: List all cycles
      responses:
        "200":
          description: List of cycles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cycle"
    post:
      operationId: createCycle
      tags: [cycles]
      summary: Create a cycle (editor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCycleBody"
      responses:
        "201":
          description: Created cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cycle"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /cycles/{id}:
    patch:
      operationId: updateCycle
      tags: [cycles]
      summary: Update a cycle (editor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCycleBody"
      responses:
        "200":
          description: Updated cycle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cycle"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCycle
      tags: [cycles]
      summary: Delete a cycle (editor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List all business goals
      responses:
        "200":
          description: List of business goals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Create a business goal (editor only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGoalBody"
      responses:
        "201":
          description: Created goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a business goal (editor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGoalBody"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a business goal (editor only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Cycle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - startDate
        - endDate
        - createdAt
    CreateCycleBody:
      type: object
      properties:
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
      required:
        - name
        - startDate
        - endDate
    UpdateCycleBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - color
        - createdAt
    CreateGoalBody:
      type: object
      properties:
        name:
          type: string
        color:
          type: string
      required:
        - name
        - color
    UpdateGoalBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
...